ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
joblib==1.3.2
scipy==1.11.4






